mongo-api/requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib
CORSMiddleware